Add Storybook coverage for `MapboxMap` and update Component styling. (#14)
This PR ports over the styling for the `MapboxMap` Component. There was very little styling to be ported over. The majority was too specific to Consulting. I also added Storybook coverage for the Component in this PR. I also changed Storybook so that it's built using Webpack instead of Vite. I did this for two reasons. One is consistency with `@yext/search-ui-react` and the second is that replacing env vars with GH secrets is very difficult to get right in Vite. Ultimately, we do want to build the Site with Vite, it's got a lot of benefits, including speed. I will create a couple Code Maintenance items. One to figure out the env var replacement in Vite and another to cut-over all our Storybook sites to be built with Vite. 

J=SLAP-2512
TEST=auto