Support a CommonJS build of the `Markdown` Component.
Previously, the `Markdown` Component did not work with CommonJS. This
is because the versions of `react-markdown`, `remark-gfm`, etc. were
ESM-only. In order to have a CommonJS-compatible Component, I needed to
downgrade these and other libraries used by `Markdown`.

Initially, the decision was made intentionally to only have an ESM-compatible
`Markdown` Component. However, that's proven troublesome for PagesJS users.
Now that there is a CommonJS variant of `Markdown`, I was able to add Jest
coverage. The Jest tests were essentially ported from the tests KG had in Alpha.

The older version of `rehype-sanitize` does not support user-supplied Schema.
That's okay, however. Product and I decided this was not an important feature
after all.

TEST=auto,manual

Ran all unit tests. Manually ensured that HTML was getting sanitized properly.