Fixed styling for numbered list (#28)
This PR is to fix this Tech Ops https://yexttest.atlassian.net/browse/TECHOPS-9091

The definition and declaration of using Lexical library to display the Rich Text content should be consistent in between the usage on entity storm project under alpha, and the usage in the React-component Github repo. In this PR, we are fixing the style difference that caused the issue on formatting nested numbered list.